Added `Bit` enum and base-2 conversions for `UnsignedInteger`.
Includes tests.

While Swift already provides the initializer `String(_:radix:)` which does a similar job, this has the disadvantages that (a) it requires the integer type to be of fixed width, which is never the case in cryptographic algorithms, and (b) the result is an ordinary String, so the type does not guarantee that the object is a valid representation of an integer (only the method).