Changed ranges from which the prime factors of the public key are chosen.
The previous approach had a serious problem! By choosing the first prime uniformly at random from 2 to `RSAKeys.publicKeyUpperBound`, there was a huge probability that it would be very large. For example, there was about 50% probability that it would be at least `RSAKeys.publicKeyUpperBound / 2`. This means that the other prime factor was restricted to very small values (in the previous example, the other factor would be forced to be 2). Most of the time the second prime factor ended up being 2, 3 or 5.

In the new version, we force the first prime factor to be the smallest one, and use the fact that the smallest prime factor of the public key must be below the square root of `RSAKeys.publicKeyUpperBound`. The tests show that both prime factors end up being very large.