Got rid of type `Prime`.
This type was too strict. It is obviously important for the theoretical correctness of the algorithms and objects, but if we just focus on the type-safety guarantees we need in the code to avoid crashes and the need to throw and handle exceptions, it is sufficient to replace `Prime` with simply `GreaterThanOne`.

This is much better now, because `GreaterThanOne` is obviously very quick to initialize. With `Prime`, if we implemented JSON encoding and decoding (so that keys can be stored in the memory), we would need to perform two primality checks when decoding. This is fast in this toy version, but unrealistic in serious implementations. By using `GreaterThanOne`, we open up the possibility that corrupt or malicious data may initialize keys with non-prime factors (so the user should be careful not to let anyone modify their stored keys), but we still have the guarantees that, e.g. we will not try to divide by 0 and so on. The code remains safe, even though the algorithm conceptually does not. The advantage is that JSON decoding of keys becomes very fast, even in a real-world application involving huge numbers.