Global improvement of access control.
By testing the framework using a playground, made public several types, members and methods that should be and were not, and made internal or private several that shouldn't be public but were. The public interface was reduced as much as possible, leaving only the actual high level encryption and decryption code accessible. Also, the keyword `internal` was being used only sometimes, while other times it was implicit, and this was changed for consistency (now it is never used, i.e. is always implicit).